requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
magic-filter==1.0.11
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2==2.9.7
pydantic==1.10.12
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
yarl==1.9.2
